Outlining initial structure to the page
Outlining general structure: 4 columns (2 sidebars, 2 central columns). Central column on the left will be a title column with the title and subtitle of a chapter (page). Center column on the right will have the contents of the chapter (page). 
The left sidebar will be a navigation bar between chapters. Don't know what the right bar will do yet :)